chore: separate utils from file_paths logic
I created two files in the utils folder, I was getting import errors and tried to give it some order, but It is provisional. I would like to have a separate class for profiling, that returns the results as numpy/pandas arrays so we can apply some filtering to them.