chore: compute scaling type for stage.
Adds the compute scaling for the performance model. Some stages don't scale based on CPUs (vertically), and only horizontally, then the compute_scaling argument would be set to "workers".
If the stage can scale vertically, then setting the compute_scaling to cpu will yield better results.
What this argument does is modify the key used in the plots as well as in the model, so if a stage can only scale horizontally, then the x axis will be workers.